// clients/admin-ui/src/types/api/models/Page_Asset_.ts
// This file is auto-generated by @hey-api/openapi-ts

import type { Asset } from "./Asset";

/**
 * Page[Asset]
 */
export type Page_Asset_ = {
  /**
   * Items
   */
  items: Array<Asset>;
  /**
   * Total
   */
  total: number;
  /**
   * Page
   */
  page: number;
  /**
   * Size
   */
  size: number;
  /**
   * Pages
   */
  pages: number;
};